Deletion controls

+++ dev/newController.py
@@ -49,8 +49,8 @@ def loadConfigs():
 	del_l = buildDeleteList(del_d,'schedconfig')
 	# Information regarding 
 	if len(del_d): emailError('Deleting queues: %s' % ','.join(del_d.keys()))
-	if len(del_d)/len(collapseDict(agisd)) >= maxDeletedQueuePercentage/100:
-		msg = 'Deleting too many queues: %d percent is higher than the maximum allowed of %d percent' % (int(len(del_d)/len(collapseDict(agisd))*100),maxDeletedQueuePercentage)
+	if float(len(del_d))/float(len(collapseDict(agisd))) >= float(maxDeletedQueuePercentage/100):
+		msg = 'Deleting too many queues: %d percent is higher than the maximum allowed of %d percent' % (int(float(len(del_d))/float(len(collapseDict(agisd)))*100),maxDeletedQueuePercentage)
 		emailError(msg)
 		print msg
 		return 1
